+++ src/components/Terminal/TerminalSettings.tsx
@@ -4,12 +4,13 @@
  * SPDX-License-Identifier: LicenseRef-Nordic-4-Clause
  */
 
-import React, { useEffect, useRef } from 'react';
+import React, { useEffect, useRef, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import {
     CollapsibleGroup,
     Dropdown,
     getPersistedTerminalSettings,
+    NumberInlineInput,
     persistTerminalSettings,
     selectedDevice,
     StateSelector,
@@ -21,6 +22,7 @@ import {
     getEchoOnShell,
     getLineEnding,
     getLineMode,
+    getScrollback,
     getSerialOptions,
     getSerialPort,
     LineEnding,
@@ -28,6 +30,7 @@ import {
     setEchoOnShell,
     setLineEnding,
     setLineMode,
+    setScrollback as setActiveScrollback,
 } from '../../features/terminal/terminalSlice';
 import { convertToDropDownItems } from '../../utils/dataConstructors';
 
@@ -37,6 +40,9 @@ export default () => {
     const lastModemOpenState = useRef(false);
     const selectedSerialport = useSelector(getSerialOptions).path;
 
+    const activeScrollback = useSelector(getScrollback);
+    const [scrollback, setScrollback] = useState(activeScrollback);
+
     const clearOnSend = useSelector(getClearOnSend);
     const lineEnding = useSelector(getLineEnding);
     const echoOnShell = useSelector(getEchoOnShell);
@@ -44,6 +50,8 @@ export default () => {
     const lineMode = useSelector(getLineMode);
 
     const dispatch = useDispatch();
+
+    const isConnected = serialPort !== undefined;
 
     const lineEndings = convertToDropDownItems<string>(
         ['NONE', 'LF', 'CR', 'CRLF'],
@@ -132,7 +140,23 @@ export default () => {
     ]);
 
     return (
-        <CollapsibleGroup heading="Terminal Settings">
+        <CollapsibleGroup heading="Terminal Settings" defaultCollapsed={false}>
+            <div
+                title="Set the number of lines it is possible to scroll in the Terminal"
+                className="tw-flex tw-justify-between"
+            >
+                Scrollback
+                <NumberInlineInput
+                    value={scrollback}
+                    onChange={setScrollback}
+                    onChangeComplete={() => {
+                        if (scrollback !== activeScrollback) {
+                            dispatch(setActiveScrollback(scrollback));
+                        }
+                    }}
+                    range={{ min: 1, max: 2 ** 64 - 1 }}
+                />
+            </div>
             <StateSelector
                 items={lineModeItems}
                 onSelect={value => dispatch(setLineMode(value === 0))}