+++ Changelog.md
@@ -3,6 +3,8 @@
 ### Added
 
 -   Custom baud rate support.
+-   Option to edit the number of lines that can be scrolled back. The default
+    `scrollback` is set to 1000 lines.
 
 ### Changed